serialize requests to avoid token not being available + hide admin
Fetching on initial component load does not work if the auth
state is only available later. Form-2 initialization logic
will not be re-run. We can't subscribe on state changes either,
so the update logic outside reagent re-rendering will have to
use something outside the reagent state => no single source of
truth. This probably is the use case for re-fetch.